Adapt for multi-user support
There are two main changes:

Keeping track of which 'encode' request a 'knn' request belongs to:

- The Sent2Vec server returns an object {id: "xxx"} as the result of the
'encode' request.
- The client (webserver) sends this ID back to the browser when
answering the browser's POST request
- The browser saves this ID in the window.sessionStorage.
- Whenever the user initiates a 'knn' request, the browser submits this
ID along with the POST request to the webserver.
- The webserver includes this ID in the params of the JSON-RPC message
to the Sent2Vec server.
- The Sent2Vec server now knows which encoded text to apply this kNN
query to.

Keeping track to which request message a response message belongs to. This
uses the required 'id' field in JSON-RPC messages:

- Before sending any type of message, the webserver generates an ID, and
includes it in the 'id' field of the JSON-RPC message to send.
- The webserver then adds an entry, indexed with this ID, to the list of
messages that it has sent, but not yet received a resonse for.
- In this entry, the webserver includes the request type of this message
('encode' or 'knn') and a handle to send a response back to the browser.
- Whenever the webserver receives a message from the Sent2Vec server, it extracts the message's JSON-RPC 'id' field and looks it up in the list of messages that it has not yet received a response for.
- The webserver now knows what type of response this is (encode or knn),
and can handle it appropriately.