Disable timeout for Ajax requests
Was 30 seconds before (in Chrome). So when the encoding of a text took
longer than this, then the Ajax request was cancelled, even though the
Sent2Vec server was still working on it.

EDIT:

The interruption of the request after 30 seconds is not caused by
the Ajax timeout, but by Heroku. Heroku kills requests that take more
than 30 sedonds, returns an HTTP 503 status code (Service Unavailable),
and issues an H12 error.

It's not possible to modify this timeout, no request to a web process
may take longer than 30 seconds.

See here:
- https://devcenter.heroku.com/articles/request-timeout
- https://devcenter.heroku.com/articles/error-codes#h12-request-timeout