Remove git-secret
It was used to encrypt the Cloud Foundry manifest.yml file, because this
file contains a RabbitMQ URI (which includes username and password of a
RabbitMQ instance).

When using Jenkins, git-secret is probably not a good choice. Because
then, when Jenkins does the Git checkout, the file will be encrypted,
and there must be a step in the pipeline which decrypts this file before
proceeding. This means, that git-secret, GnuPG, and a key pair that has
been added to the authorised keys of git-secret, must be installed on
the Jenkins build server on which the pipeline is run.

This is probably hard to achieve, given that pipelines can also run on
slave nodes.

So, a better option to hide credentials from a public Git repository
might be to use the Jenkins Credentials and Credentials Binding plugin.
It allows to define credentias in the Jenkins GUI, and then to inject
them as variables in the pipeline scripts.