In Deploy to Dev, push app to Cloud Foundry
cf push reads the parameters from manifest.yml.

manifest.yml does not contain the RABBITMQ_URI environment variable,
that is necessary for the Docker image, because this URI contains
username and password.

Therefore, after cf push, we call cf set-env and set this environment
variable through a Jenkins credential.

cf set-env comes after cf push, because, if it would be the other way
around and the app is deployed for the very first time, the app wouldn't
yet exist at the time cf set-env is executed (because only cf push
creates the app if it doesn't yet exist).